code/requirements.txt
astropy==4.3.post1
attrs==20.3.0
cycler==0.11.0
Cython==0.29.21
fonttools==4.28.1
h5py==3.3.0
hypothesis==6.0.0
iniconfig==1.1.1
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pluggy==0.13.1
py==1.10.0
pybind11==2.6.1
pyerfa==2.0.0
pygadmpa @ file:///users/arawling/tools/pygad
pyparsing==3.0.6
pytest==6.2.1
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.2
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sortedcontainers==2.3.0
toml==0.10.2
tomli==1.2.2
